pyproject.toml
[project]
name = "fastpaip-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "GuillermoLB", email = "guillelb99@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "openai>=2.6.0",
    "plummy",
    "pydantic-settings>=2.11.0",
]

[project.scripts]
fastpaip-app = "fastpaip_app:main"

[build-system]
requires = ["uv_build>=0.9.0,<0.10.0"]
build-backend = "uv_build"

[tool.uv.sources]
plummy = { path = "./plummy" }
